# Don't save exact duplicates in merged index