@aawsome 2020-08-04
There is another issue I just came around when thinking about this change once again: The packIDs are not deduplicated so far.. 
I need to think about how to best do this and change this to WIP in meanwhile..