@aawsome 2020-08-04
I changed this such that now the check is performed in `Index.merge()`. IndexMap` stays unchanged now.