@aawsome 2020-08-05
> @aawsome Did you check the masterIndex benchmarks regarding the performance impact of the additional `foreach` loop?

Note that the additional `foreachID` loop is `O(1)` itself (at least if the undelying hash is not degenerated). Hence the total complexity of the merge is still roughly (neglecting hash growing of `idx`) `O(n)` with `n` being the number of elements of the index to be merged, i.e. `idx2`. 

Moreover this change only affects the merge itself, and in real-world-scenarios this occurs during 

- loading the index from index files into memory (should be dominated by loading and extracting the files)
- saving a new finished index during backup (this should happen quite rarely as backup itself should be mainly dominated by 

So I do not expect any impact on real-life usage.

About benchmarking: Seems to me that none of the current benchmarks actually includes merging of indexes in the measurements.
I added the benchmark `MasterIndexAlloc` which creates Indexes and merges them under measurement.
Here are the results comparing 5e63294 with b112533:
```
name                old time/op    new time/op    delta
MasterIndexAlloc-4     334ms ± 1%     460ms ± 1%  +37.68%  (p=0.000 n=9+8)

name                old alloc/op   new alloc/op   delta
MasterIndexAlloc-4     258MB ± 0%     258MB ± 0%     ~     (p=0.101 n=10+8)

name                old allocs/op  new allocs/op  delta
MasterIndexAlloc-4      317k ± 0%      317k ± 0%     ~     (p=0.267 n=10+6)
```